Admin: Fix BasePopupChoiceWidget filter
* Move flatatt_filter to shoop/admin/utils/forms.py
* Use flatatt_filter to skip attrs which value is not set

Since (d53b733) using data-filter="None" causes errors in popups.
